first commit to develop...
We have a fundamental idea of how this is going to go....
We have a controller which:
- listens for SLAM stuff
- requests new paths on SLAM stuff arrival
- requests for objects to be tracked on SLAM stuff arrival
- publishes most recently avaiilable path

see the readme, for a link to the flow chart

the current state is:
There exists a waypoint follower which can follow waypoints.
A controller which does talk to path planner, and
a path planner that can respond.

We're a dedicated interface package with custom messages and services. All is currently writting in
 python, but not against trying to write it in C++.

 Welcome to the world 'Novo'.